moving files around and delete all button complete
