+++ frontend/src/components/css/Notes.css
