+++ frontend/src/components/tests/Update.test.js
@@ -1,6 +1,6 @@
 import '@testing-library/jest-dom';
 import { render, screen, fireEvent } from '@testing-library/react';
-import Notes from './Notes';
+import Notes from '../Notes';
 
 test('Update button exists and updates parent state', () => {
   const onUpdateMock = jest.fn();
